significant internal refactor for modularity
broke the logic around data loading and splitting into separate functions and classes instead of holding everything under fastprop
