+++ fastprop/__init__.py
@@ -1,8 +1,3 @@
 from .defaults import DEFAULT_TRAINING_CONFIG
-from .hopt import hopt_fastprop
-from .predict import predict_fastprop
-from .preprocessing import preprocess
-from .shap import shap_fastprop
-from .train import train_fastprop
 
-__all__ = ["DEFAULT_TRAINING_CONFIG", "hopt_fastprop", "predict_fastprop", "preprocess", "shap_fastprop", "train_fastprop"]
+__all__ = ["DEFAULT_TRAINING_CONFIG"]
