--- fastprop/utils/__init__.py
@@ -1,17 +1,0 @@
-from .calculate_descriptors import _get_descs, calculate_mordred_desciptors
-from .descriptor_lists import ALL_2D, SUBSET_947, descriptors_lookup
-from .load_data import load_from_csv, load_saved_desc
-from .select_descriptors import mordred_descriptors_from_strings
-from .validate_config import validate_config
-
-__all__ = [
-    "_get_descs",
-    "calculate_mordred_desciptors",
-    "ALL_2D",
-    "SUBSET_947",
-    "descriptors_lookup",
-    "load_from_csv",
-    "load_saved_desc",
-    "mordred_descriptors_from_strings",
-    "validate_config",
-]
