--- fastprop/utils/validate_config.py
@@ -1,7 +1,0 @@
-from fastprop.defaults import init_logger
-
-logger = init_logger(__name__)
-
-
-def validate_config(config: dict):
-    logger.warn("TODO: Input validation has not been implemented! Skipping.")
